% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVClassify.R
\name{BVClassify}
\alias{BVClassify}
\title{Predict BV classes from a given dataframe}
\usage{
BVClassify(dataset, plot_HM = FALSE)
}
\arguments{
\item{dataset}{dataframe: The columns correspond to the species used for classification and the rows correspond to the samples.}

\item{plot_HM}{boolean:TRUE for make heatmap and FALSE for not}
}
\value{
dataframe: the first three columns correspond to the probability of belonging to each of the classes and the last one to the label of the class you belong to. Or list with dataframe and heatmap plot.
}
\description{
Predict BV classes from a given dataframe
}
\examples{
BVClassify(normdataset(sample_data))
}
