% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ncbi_ID.R
\name{convert_ncbi_ID}
\alias{convert_ncbi_ID}
\title{Convert NCBI IDs on Species names}
\usage{
convert_ncbi_ID(dataset)
}
\arguments{
\item{dataset}{dataframe: The columns correspond to the species raw counts used for classification and the rows correspond to the samples}
}
\value{
dataframe: same dataframe but with column names changed
}
\description{
Convert NCBI IDs on Species names
}
