% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_annotation-data.R
\docType{data}
\name{df_annotation}
\alias{df_annotation}
\title{Dataframe with NCBI references for the 8 species of the model}
\format{
A reference dataframe
}
\usage{
data(df_annotation)
}
\description{
Dataframe with NCBI references for the 8 species of the model
}
\keyword{annotation}
