% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelC-data.R
\docType{data}
\name{modelC}
\alias{modelC}
\title{Pre-trained Random Forest Model to classify BV patients}
\format{
A wrapped learner object
}
\usage{
data(modelC)
}
\description{
Pre-trained Random Forest Model to classify BV patients
}
\keyword{models}
