% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normdataset.R
\name{normdataset}
\alias{normdataset}
\title{Make log2 Normalization of a given dataframe}
\usage{
normdataset(dataset)
}
\arguments{
\item{dataset}{dataframe: The columns correspond to the species raw counts used for classification and the rows correspond to the samples}
}
\value{
dataframe: Return a dataframe with the same structure but with log2 normalised counts
}
\description{
Make log2 Normalization of a given dataframe
}
\examples{
normdataset(iris)
}
