% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data-data.r
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Demo datasets for use functions}
\format{
A datadframe with species in columns and samples in rows
}
\usage{
data(sample_data)
}
\description{
Demo datasets for use functions
}
\keyword{data}
